dao compiles with groups table

+++ blogdown-model/src/lib.rs
@@ -1,12 +1,12 @@
 pub struct Group 
 {
-    pub id: u64,
+    pub id: u32,
     pub name: String,
 }
 
 pub struct User<'a>
 {
-    pub id: u64,
+    pub id: u32,
     pub name: String,
     pub email: String,
     pub groups: Vec<Group>,
@@ -16,14 +16,14 @@ pub struct User<'a>
 
 pub struct Comment<'a> 
 {
-    pub id: u64,
+    pub id: u32,
     pub content: String,
     pub reply_to: Option<&'a Comment<'a>>,
 }
 
 pub struct Blog<'a>
 {
-    pub id: u64,
+    pub id: u32,
     pub title: String,
     pub content: String,
     pub comments: Vec<Comment<'a>>,
